Auth with Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -71,6 +71,9 @@ class QuestionsController extends Controller
     public function edit(Question $question)
     {
         // $question = Question::findOrFail($id);
+        if(\Gate::denies('update-question',$question)){
+            abort(403 ,"Access Denied !");
+        }
         return view('questions.edit' ,compact('question'));
     }
 
@@ -83,6 +86,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+
+        if(\Gate::denies('update-question',$question)){
+            abort(403 ,"Access Denied !");
+        }
+
         $question->update($request->only('title','body'));
 
         return redirect('/questions')->with('success', 'Your question has been updated');
@@ -96,6 +104,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question)){
+            abort(403 ,"Access Denied !");
+        }
         $question->delete();
         return redirect('/questions')->with('success','Deleted!!');
     }
